backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
Faker==37.1.0
pillow==11.1.0
psycopg2==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
